#1339 removed scala shell commands
